requirements.txt
streamlit==1.30.0
supabase==2.4.0
gotrue==2.4.2
httpx==0.24.1
pandas==2.1.4
python-dotenv==1.0.1
openpyxl==3.1.2
